+++ start/src/App.vue
@@ -1,5 +1,19 @@
 <script lang="ts" setup></script>
 
-<template></template>
+<template>
+  <f-alert bold close type="default">默认提示信息</f-alert>
+  <f-alert bold close type="primary">主要提示信息</f-alert>
+  <f-alert bold close type="success">成功提示信息</f-alert>
+  <f-alert bold close type="danger">危险提示信息</f-alert>
+  <f-alert bold close type="warning">警告提示信息</f-alert>
+  <f-alert bold close type="info">信息提示信息</f-alert>
+
+  <!-- <f-alert simple type="default">默认提示信息</f-alert>
+  <f-alert simple type="primary">主要提示信息</f-alert>
+  <f-alert simple type="success">成功提示信息</f-alert>
+  <f-alert simple type="danger">危险提示信息</f-alert>
+  <f-alert simple type="warning">警告提示信息</f-alert>
+  <f-alert simple type="info">信息提示信息</f-alert> -->
+</template>
 
 <style lang="scss" scoped></style>